+++ src/routes/graphql/points/profiles/typeProfile.ts
